Mui5 migration: Breaking changes in v5: styles and theme (#514)
