+++ packages/react-ui/src/widgets/legend/LegendCategories.js
@@ -68,10 +68,7 @@ export default LegendCategories;
 // Aux
 const useStyles = makeStyles((theme) => ({
   legendCategories: {
-    alignItems: 'center',
-    '&:hover': {
-      '& $circle': {}
-    }
+    alignItems: 'center'
   },
   marker: {
     whiteSpace: 'nowrap',
