+++ packages/react-ui/storybook/stories/common/Text-field.stories.js
@@ -20,11 +20,25 @@ const options = {
       }
     },
     required: {
+      defaultValue: false,
       control: {
         type: 'boolean'
       }
     },
     disabled: {
+      defaultValue: false,
+      control: {
+        type: 'boolean'
+      }
+    },
+    error: {
+      defaultValue: false,
+      control: {
+        type: 'boolean'
+      }
+    },
+    multiline: {
+      defaultValue: false,
       control: {
         type: 'boolean'
       }
@@ -32,6 +46,12 @@ const options = {
     label: {
       control: {
         type: 'text'
+      }
+    },
+    margin: {
+      control: {
+        type: 'select',
+        options: ['dense', 'none', 'normal']
       }
     }
   },
